+++ app/routes/sketch+/artboards+/$slug_+/services/layer/design/move-down.service.ts
@@ -16,7 +16,7 @@ export const layerDesignMoveDownService = async ({
 	updateSelectedDesignId?: IDesignIdOrNull
 }) => {
 	try {
-		console.log('layerDesignDeleteService')
+		console.log('layer move down')
 		const updateSelectedDesignStrategy = new LayerUpdateSelectedDesignStrategy()
 		return designMoveDownService({
 			userId,
